+++ demo/src/main.ts
@@ -1,4 +1,5 @@
 import { createApp } from "vue";
+import "lenis/dist/lenis.css";
 import App from "./App.vue";
 import { router } from "./router";